requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
pytest==8.3.4
pytest-cov==6.0.0
PyYAML==6.0.2
